feat(various): Now pushing the user to the login page when their account has been deleted
Realized that the AccountChecker and AccountProblemPage weren't necessary. So, both of those files have been deleted.
